Moved geometry transformations from imgproc to 3d, future geometry module.
