+++ modules/3d/src/min_enclosing_convex_polygon.cpp
