+++ modules/3d/src/min_enclosing_triangle.cpp
