+++ modules/3d/src/precomp.hpp
@@ -82,6 +82,7 @@
 #include <unordered_set>
 #include <map>
 #include <unordered_map>
+#include <cmath>
 
 #define GET_OPTIMIZED(func) (func)
