+++ modules/3d/src/rotcalipers.cpp
