+++ modules/3d/test/test_convhull.cpp
@@ -598,6 +598,20 @@ TEST(Imgproc_minAreaRect, roundtrip_accuracy)
     EXPECT_LT(std::abs(rect.angle - rect_out.angle), 1e-5);
 }
 
+TEST(Imgproc_PointPolygonTest, regression_10222)
+{
+    vector<Point> contour;
+    contour.push_back(Point(0, 0));
+    contour.push_back(Point(0, 100000));
+    contour.push_back(Point(100000, 100000));
+    contour.push_back(Point(100000, 50000));
+    contour.push_back(Point(100000, 0));
+
+    const Point2f point(40000, 40000);
+    const double result = cv::pointPolygonTest(contour, point, false);
+    EXPECT_GT(result, 0) << "Desired result: point is inside polygon - actual result: point is not inside polygon";
+}
+
 TEST(Imgproc_minEnclosingTriangle, regression_17585)
 {
     const int N = 3;
