+++ modules/dnn/src/model.cpp
@@ -10,6 +10,7 @@
 #include <iterator>
 
 #include <opencv2/imgproc.hpp>
+#include <opencv2/3d.hpp>
 
 namespace cv {
 namespace dnn {
