+++ modules/dnn/test/test_model.cpp
@@ -4,6 +4,7 @@
 
 #include "test_precomp.hpp"
 #include <opencv2/dnn/shape_utils.hpp>
+#include <opencv2/3d.hpp>
 #include "npy_blob.hpp"
 #include <map>
 #include <set>
