+++ modules/features/src/precomp.hpp
@@ -45,6 +45,7 @@
 
 #include "opencv2/features.hpp"
 #include "opencv2/imgproc.hpp"
+#include "opencv2/3d.hpp"
 
 #include "opencv2/core/utility.hpp"
 #include "opencv2/core/private.hpp"
