+++ samples/cpp/geometry.cpp
@@ -16,6 +16,7 @@
  *
  *********************************************************************************/
 
+#include "opencv2/3d.hpp"
 #include "opencv2/imgproc.hpp"
 #include "opencv2/highgui.hpp"
 #include <iostream>
