+++ samples/cpp/snippets/intersectExample.cpp
@@ -4,6 +4,7 @@
  * A program that illustrates intersectConvexConvex in various scenarios
  */
 
+#include "opencv2/3d.hpp"
 #include "opencv2/imgproc.hpp"
 #include "opencv2/highgui.hpp"
