+++ samples/cpp/snippets/lsd_lines.cpp
@@ -1,3 +1,4 @@
+#include "opencv2/3d.hpp"
 #include "opencv2/imgproc.hpp"
 #include "opencv2/imgcodecs.hpp"
 #include "opencv2/highgui.hpp"
