+++ samples/cpp/snippets/squares.cpp
@@ -4,6 +4,7 @@
 // each image
 
 #include "opencv2/core.hpp"
+#include "opencv2/3d.hpp"
 #include "opencv2/imgproc.hpp"
 #include "opencv2/imgcodecs.hpp"
 #include "opencv2/highgui.hpp"
