+++ samples/cpp/tutorial_code/ShapeDescriptors/hull_demo.cpp
@@ -6,6 +6,7 @@
 
 #include "opencv2/imgcodecs.hpp"
 #include "opencv2/highgui.hpp"
+#include "opencv2/3d.hpp"
 #include "opencv2/imgproc.hpp"
 #include <iostream>
