+++ samples/cpp/tutorial_code/ShapeDescriptors/pointPolygonTest_demo.cpp
@@ -5,6 +5,7 @@
  */
 
 #include "opencv2/highgui.hpp"
+#include "opencv2/3d.hpp"
 #include "opencv2/imgproc.hpp"
 #include <iostream>
